+++ weaver/app.py
@@ -10,6 +10,7 @@ import yaml
 from pyramid.config import Configurator
 
 from weaver.config import WEAVER_DEFAULT_REQUEST_OPTIONS_CONFIG, get_weaver_config_file, get_weaver_configuration
+from weaver.database import get_db
 from weaver.processes.builtin import register_builtin_processes
 from weaver.processes.utils import register_wps_processes_from_config
 from weaver.utils import get_settings, parse_extra_options, setup_cache, setup_loggers
@@ -57,6 +58,10 @@ def main(global_config, **settings):
         local_config.include("pyramid_celery")
         local_config.configure_celery(global_config["__file__"])
     local_config.include("weaver")
+
+    LOGGER.info("Running database migration...")
+    db = get_db(settings)
+    db.run_migration()
 
     if settings.get("weaver.celery", False):
         LOGGER.info("Celery runner detected. Skipping process registration.")