+++ weaver/store/base.py
@@ -4,10 +4,12 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     import datetime
     from typing import Any, Dict, List, Optional, Tuple, Union
+
     from pyramid.request import Request
     from pywps import Process as ProcessWPS
+
     from weaver.datatype import Bill, Job, Process, Quote, Service
-    from weaver.typedefs import DatetimeIntervalType, TypedDict
+    from weaver.typedefs import AnyUUID, DatetimeIntervalType, TypedDict
 
     JobGroupCategory = TypedDict("JobGroupCategory",
                                  {"category": Dict[str, Optional[str]], "count": int, "jobs": List[Job]})
@@ -129,7 +131,7 @@ class StoreJobs(StoreInterface):
 
     @abc.abstractmethod
     def fetch_by_id(self, job_id):
-        # type: (str) -> Job
+        # type: (AnyUUID) -> Job
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -173,7 +175,7 @@ class StoreQuotes(StoreInterface):
 
     @abc.abstractmethod
     def fetch_by_id(self, quote_id):
-        # type: (str) -> Quote
+        # type: (AnyUUID) -> Quote
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -197,7 +199,7 @@ class StoreBills(StoreInterface):
 
     @abc.abstractmethod
     def fetch_by_id(self, bill_id):
-        # type: (str) -> Bill
+        # type: (AnyUUID) -> Bill
         raise NotImplementedError
 
     @abc.abstractmethod